integrated auth check in edit and delete

+++ controller/postController.js
@@ -1,6 +1,7 @@
 const Post = require('../models/post')
 const getUserDetails = require('../middleware/userDetails')
 const deleteImage = require("../middleware/deletePhoto")
+const jwtCheck = require('../middleware/auth')
 
 const getPosts = async (req, res) => {
     const post = await Post.find();
@@ -10,7 +11,7 @@ const getPosts = async (req, res) => {
     res.json(post);
 }
 const createPosts = async (req, res) => {
-    if (!req?.headers?.authorization) {
+    if (!jwtCheck) {
         return res.status(401).json({ message: "Unauthorized" })
     }
     if (!req?.body?.title || !req?.body?.datetime || !req?.body?.body) {
@@ -35,7 +36,6 @@ const createPosts = async (req, res) => {
                 sub: sub.split('|')[1]
             }
         })
-
         const _id = response._id.toHexString();
         const data = response._doc.user
         res.status(200).json({ _id, user:data});
@@ -46,33 +46,51 @@ const createPosts = async (req, res) => {
 }
 
 const updatePosts = async (req, res) => {
+    if(!jwtCheck){
+        return res.status(401).json({ message: "Unauthorized" })
+    }
+    const accessToken = req.headers?.authorization?.split(' ')[1];
+    const user = await getUserDetails(accessToken)
+    if (req.body?.user?.sub != user?.data?.sub?.split('|')[1]){
+        return res.status(401).json({ message: "Unauthorized" })
+    }
     if (!req?.params?.id) {
         return res.status(400).json({ message: "Not sufficient information provided" });
     }
     const post = await Post.findOne({ _id: req.params.id }).exec();
     if (!post) {
-        return res.status(204).json({ message: `NO such Post found` });
+        return res.status(204).json({ message: `No such post found` });
     }
     const { title, body, datetime, image, imgId } = req.body;
     post.body = body;
     post.title = title;
-    post.datetime = datetime
+    post.datetime = datetime;
+
     if (image) {
         await deleteImage(post.imgId);
         post.image = image
         post.imgId = imgId
     }
+
     const result = await post.save();
     res.json(result);
 }
 const deletePosts = async (req, res) => {
+    if(!jwtCheck){
+        return res.status(401).json({ message: "Unauthorized" })
+    }
+    const accessToken = req.headers?.authorization?.split(' ')[1];
+    const user = await getUserDetails(accessToken)
+    if (req.body?.user?.sub != user?.data?.sub?.split('|')[1]){
+        return res.status(401).json({ message: "Unauthorized" })
+    }
     if (!req?.params?.id) {
         return res.status(400).json({ message: "Not sufficient information provided" });
     }
     const post = await Post.findOne({ _id: req.params.id }).exec();
     await deleteImage(post.imgId);
     if (!post) {
-        return res.status(204).json({ message: `NO such Post found` });
+        return res.status(204).json({ message: `No such Post found` });
     }
     const result = await post.deleteOne({ _id: req.body.id });
     res.json(result);
